requirements.txt
git+https://github.com/iksnagreb/brevitas.git@34c350afde0a69b7a39b13eb4748b979bdae1eff
git+https://github.com/iksnagreb/qonnx.git@b6a23ba22dd25d46b462145e382c04237c9a6952
git+https://github.com/eki-project/finn-plus.git@a2e6a2f562a380463f6fb03e41eba63e1597e66f
numpy~=1.26.4
onnxoptimizer~=0.3.13
onnxruntime~=1.17.3
onnx~=1.17.0
PyYAML~=6.0.2
torch~=2.5.0
pandas~=2.2.3
dvc[webdav]~=3.55.2
tqdm~=4.66.5
h5py~=3.12.1
scikit-learn~=1.5.2
pyarrow~=15.0.0
